# book_back/src/main/java/com/example/book_back/feedback/FeedbackResponse.java
package com.example.book_back.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;
}